backend/foodgram/requirements.txt
Django==3.2.16
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==3.13.0
djoser==2.2.0
oauthlib==3.2.2
Pillow==10.0.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
python-dotenv==1.0.0